requirements.txt
django==3.1.1
djangorestframework>=3.11.1,<3.12.0
django-filter==2.3.0
djangorestframework-jwt>=1.11.0,<1.12.0
django-import-export>=2.4.0,<2.5.0
django-resized>=0.3.11,<0.4.0
django-modeltranslation >= 0.16.2

channels>=2.4.0,<2.5.0
channels-redis>=3.1.0,<3.2.0
psycopg2>=2.8.6,<2.9.0
pillow==8.0.1
sorl-thumbnail
gunicorn==20.0.4
django-heroku
whitenoise
django-admin-rangefilter
django-admin-inline-paginator
django-cors-headers>=3.5.0,<3.6.0

celery
redis
whitenoise
